Extend native Object prototype with EventEmitter properties
This allow the use of direct events to be used instead of only allowing global
events. The change that needed to be made for this to be possible was extending
Object.prototype with EE's properties but making them not enumerable. Doing this
previously with enumerable properties caused marshalling JS Objects to
NSDictionaries to include those enumerable properties (this was found by
investigating the JSValue toDictionary method in the JSC API source in WebKit).