Add headers as an option in the existing API
Allow users to pass either a URL string or an object with the string and
optional headers. Updated tests and changed them to use the JSC bridge
API instead of just evaluating code. This makes it a lot easier to debug
what might be going wrong when a test fails.